+++ cpp/brain/controller/LayeredExtCPPN.h
@@ -50,9 +50,9 @@ namespace revolve
       // vector containing all neurons
       std::vector< std::vector< NeuronPtr>> layers_;
       // positions for indexing into the outputs_ buffer for each output neuron
-      std::map< NeuronPtr, int > outputPositionMap_;
+      std::map< NeuronPtr, size_t > outputPositionMap_;
       // positions for indexing into the inputs_ buffer for each input neuron
-      std::map< NeuronPtr, int > inputPositionMap_;
+      std::map< NeuronPtr, size_t > inputPositionMap_;
       // Map neuron id strings to Neuron objects
       std::map< std::string, NeuronPtr > idToNeuron_;
       // vector of all the neural connections
@@ -119,11 +119,11 @@ namespace revolve
 
       /// \brief positions for indexing into the outputs_ buffer for each
       /// output neuron
-      std::map< NeuronPtr, int > outputPositionMap_;
+      std::map< NeuronPtr, size_t > outputPositionMap_;
 
       /// \brief positions for indexing into the inputs_ buffer for each
       /// input neuron
-      std::map< NeuronPtr, int > inputPositionMap_;
+      std::map< NeuronPtr, size_t > inputPositionMap_;
 
       /// \brief Map neuron id strings to Neuron objects
       std::map< std::string, NeuronPtr > idToNeuron_;
